Move status field from edit-site to fields
