+++ packages/fields/src/fields/index.ts
@@ -4,3 +4,4 @@ export { default as orderField } from './order';
 export { default as featuredImageField } from './featured-image';
 export { default as parentField } from './parent';
 export { default as passwordField } from './password';
+export { default as statusField } from './status';
